# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ProductKind = typeof ProductKind[keyof typeof ProductKind];


export const ProductKind = {
  SERVICE: 'SERVICE',
  PRODUCT: 'PRODUCT',
} as const;

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  kind: ProductKind;
  price: number;
  unit: string;
  stockTracking: boolean;
  active: boolean;
}

export interface TransactionItem {
  id: number;
  productId: number;
  name: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  unit: string;
}

export type TransactionPaymentMethod = typeof TransactionPaymentMethod[keyof typeof TransactionPaymentMethod];


export const TransactionPaymentMethod = {
  CASH: 'CASH',
  QRIS: 'QRIS',
  TRANSFER: 'TRANSFER',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
} as const;

export interface Transaction {
  id: number;
  number: string;
  createdAt: string;
  cashier: string;
  items: TransactionItem[];
  total: number;
  paid: number;
  change: number;
  paymentMethod: TransactionPaymentMethod;
  status: TransactionStatus;
  /** @nullable */
  cancellationReason?: string | null;
}

export interface TransactionItemInput {
  productId: number;
  /** @exclusiveMinimum 0 */
  quantity: number;
}

export type TransactionInputPaymentMethod = typeof TransactionInputPaymentMethod[keyof typeof TransactionInputPaymentMethod];


export const TransactionInputPaymentMethod = {
  CASH: 'CASH',
  QRIS: 'QRIS',
  TRANSFER: 'TRANSFER',
} as const;

export interface TransactionInput {
  /** @minItems 1 */
  items: TransactionItemInput[];
  paymentMethod: TransactionInputPaymentMethod;
  /** @minimum 0 */
  paid: number;
}

export interface CancelTransactionInput {
  /** @minLength 3 */
  reason: string;
}

export type InventoryItemStatus = typeof InventoryItemStatus[keyof typeof InventoryItemStatus];


export const InventoryItemStatus = {
  HEALTHY: 'HEALTHY',
  LOW: 'LOW',
  OUT: 'OUT',
} as const;

export interface InventoryItem {
  id: number;
  name: string;
  sku: string;
  category: string;
  unit: string;
  currentStock: number;
  minimumStock: number;
  status: InventoryItemStatus;
}

export type PaymentBreakdownMethod = typeof PaymentBreakdownMethod[keyof typeof PaymentBreakdownMethod];


export const PaymentBreakdownMethod = {
  CASH: 'CASH',
  QRIS: 'QRIS',
  TRANSFER: 'TRANSFER',
} as const;

export interface PaymentBreakdown {
  method: PaymentBreakdownMethod;
  total: number;
}

export interface HourlyRevenue {
  hour: string;
  total: number;
}

export interface CategorySales {
  category: string;
  total: number;
}

export interface DashboardSummary {
  date: string;
  revenueToday: number;
  revenueYesterday: number;
  transactionCount: number;
  averageTransaction: number;
  expensesToday: number;
  netCash: number;
  paymentBreakdown: PaymentBreakdown[];
  hourlyRevenue: HourlyRevenue[];
  categorySales: CategorySales[];
}

export interface Activity {
  id: number;
  createdAt: string;
  actor: string;
  action: string;
  object: string;
  detail: string;
}

export type GetActivityParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type ListProductsParams = {
category?: string;
search?: string;
};

export type ListTransactionsParams = {
status?: ListTransactionsStatus;
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListTransactionsStatus = typeof ListTransactionsStatus[keyof typeof ListTransactionsStatus];


export const ListTransactionsStatus = {
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
} as const;