# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Owner overview and activity
  - name: products
    description: Products and services
  - name: transactions
    description: Sales transactions and cancellation
  - name: inventory
    description: Inventory and stock movements
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get today's shop overview
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Get recent audit activity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 8
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List active products and services
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PAID, CANCELLED]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a paid transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions/{id}/cancel:
    post:
      operationId: cancelTransaction
      tags: [transactions]
      summary: Cancel a transaction without deleting it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelTransactionInput"
      responses:
        "200":
          description: Cancelled transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid cancellation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        kind:
          type: string
          enum: [SERVICE, PRODUCT]
        price: { type: number }
        unit: { type: string }
        stockTracking: { type: boolean }
        active: { type: boolean }
      required: [id, name, sku, category, kind, price, unit, stockTracking, active]
    TransactionItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        name: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        subtotal: { type: number }
        unit: { type: string }
      required: [id, productId, name, quantity, unitPrice, subtotal, unit]
    Transaction:
      type: object
      properties:
        id: { type: integer }
        number: { type: string }
        createdAt: { type: string, format: date-time }
        cashier: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/TransactionItem" }
        total: { type: number }
        paid: { type: number }
        change: { type: number }
        paymentMethod:
          type: string
          enum: [CASH, QRIS, TRANSFER]
        status:
          type: string
          enum: [PAID, CANCELLED]
        cancellationReason:
          type: ["string", "null"]
      required: [id, number, createdAt, cashier, items, total, paid, change, paymentMethod, status]
    TransactionItemInput:
      type: object
      properties:
        productId: { type: integer }
        quantity: { type: number, exclusiveMinimum: 0 }
      required: [productId, quantity]
    TransactionInput:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/TransactionItemInput" }
        paymentMethod:
          type: string
          enum: [CASH, QRIS, TRANSFER]
        paid: { type: number, minimum: 0 }
      required: [items, paymentMethod, paid]
    CancelTransactionInput:
      type: object
      properties:
        reason: { type: string, minLength: 3 }
      required: [reason]
    InventoryItem:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        unit: { type: string }
        currentStock: { type: number }
        minimumStock: { type: number }
        status:
          type: string
          enum: [HEALTHY, LOW, OUT]
      required: [id, name, sku, category, unit, currentStock, minimumStock, status]
    DashboardSummary:
      type: object
      properties:
        date: { type: string, format: date }
        revenueToday: { type: number }
        revenueYesterday: { type: number }
        transactionCount: { type: integer }
        averageTransaction: { type: number }
        expensesToday: { type: number }
        netCash: { type: number }
        paymentBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/PaymentBreakdown"
        hourlyRevenue:
          type: array
          items:
            $ref: "#/components/schemas/HourlyRevenue"
        categorySales:
          type: array
          items:
            $ref: "#/components/schemas/CategorySales"
      required: [date, revenueToday, revenueYesterday, transactionCount, averageTransaction, expensesToday, netCash, paymentBreakdown, hourlyRevenue, categorySales]
    PaymentBreakdown:
      type: object
      properties:
        method:
          type: string
          enum: [CASH, QRIS, TRANSFER]
        total: { type: number }
      required: [method, total]
    HourlyRevenue:
      type: object
      properties:
        hour: { type: string }
        total: { type: number }
      required: [hour, total]
    CategorySales:
      type: object
      properties:
        category: { type: string }
        total: { type: number }
      required: [category, total]
    Activity:
      type: object
      properties:
        id: { type: integer }
        createdAt: { type: string, format: date-time }
        actor: { type: string }
        action: { type: string }
        object: { type: string }
        detail: { type: string }
      required: [id, createdAt, actor, action, object, detail]