# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySales } from './categorySales';
import type { HourlyRevenue } from './hourlyRevenue';
import type { PaymentBreakdown } from './paymentBreakdown';

export interface DashboardSummary {
  date: Date;
  revenueToday: number;
  revenueYesterday: number;
  transactionCount: number;
  averageTransaction: number;
  expensesToday: number;
  netCash: number;
  paymentBreakdown: PaymentBreakdown[];
  hourlyRevenue: HourlyRevenue[];
  categorySales: CategorySales[];
}