# lib/api-zod/src/generated/types/paymentBreakdown.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentBreakdownMethod } from './paymentBreakdownMethod';

export interface PaymentBreakdown {
  method: PaymentBreakdownMethod;
  total: number;
}