# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductKind } from './productKind';

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  kind: ProductKind;
  price: number;
  unit: string;
  stockTracking: boolean;
  active: boolean;
}