# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputPaymentMethod } from './transactionInputPaymentMethod';
import type { TransactionItemInput } from './transactionItemInput';

export interface TransactionInput {
  /** @minItems 1 */
  items: TransactionItemInput[];
  paymentMethod: TransactionInputPaymentMethod;
  /** @minimum 0 */
  paid: number;
}