# lib/api-zod/src/generated/types/transactionItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionItemInput {
  productId: number;
  /** @exclusiveMinimum 0 */
  quantity: number;
}